third_party/blink/web_tests/platform/generic/svg/animations/updated-attributes-expected.txt

PASS: animated width is 100
